requirements.txt
pyaudio
vosk
keyboard
arabic-reshaper
python-bidi
requests
gtts
playsound
pyttsx3